requirements.txt
py-cord>=2.6.1
python-dotenv>=0.21.0